refactor: extract to small components
